requirements.txt
streamlit
psycopg2-binary
pandas
requests
beautifulsoup4
ddgs
openpyxl
